Refactored Location services.
